+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Added documentation for the `option` decorator.
 - Added documentation for the `command` decorator.
 - Added documentation for the `env` decorator.
+- Added documentation for the `group` decorator.
 
 ### Updated