+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Added documentation for the `env` decorator.
 - Added documentation for the `group` decorator.
 - Added documentation for the `tag` decorator.
+- Added documentation for the `RootNode` class.
 
 ### Updated