+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Added documentation for the `RootNode` class.
 - Added documentation for the `ParentNode` class.
 - Added documentation for the `ChildNode` class.
+- Added documentation for the `Tree` class.
 
 ### Updated